{{cookiecutter.project_slug}}/requirements.txt
# Django
Django=={{ cookiecutter.django_version }}
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0

# Authentication
djoser==2.2.0
djangorestframework-simplejwt==5.3.0

# Security
django-axes==6.1.2

# Database
psycopg2-binary==2.9.9

# Cache & Task Queue
{% if cookiecutter.use_redis == "yes" %}
redis==5.0.1
django-redis==5.4.0
{% endif %}
{% if cookiecutter.use_celery == "yes" %}
celery==5.3.4
{% endif %}

# Development
django-debug-toolbar==4.2.0

